+++ src/main/java/zighang2/zighang/global/auth/jwt/KakaoUtil.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
-import zighang2.zighang.web.dto.KakaoDto;
+import zighang2.zighang.web.dto.kakaoLogin.KakaoDto;
 
 @Service
 @RequiredArgsConstructor
