+++ src/main/java/zighang2/zighang/web/repository/JobPostingRepository.java
@@ -2,13 +2,12 @@ package zighang2.zighang.web.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import reactor.core.publisher.Flux;
-import zighang2.zighang.web.domain.JobPosting;
+import zighang2.zighang.web.domain.JobRecommend;
 
 import java.util.List;
 
 @Repository
-public interface JobPostingRepository extends JpaRepository<JobPosting, Long> {
+public interface JobPostingRepository extends JpaRepository<JobRecommend, Long> {
 
-    List<JobPosting> findTop50ByOrderByIdDesc();
+    List<JobRecommend> findTop10ByOrderByIdDesc();
 }
